Fix image overflow
